Relocate client-facing conversation views to api-model::client (#377)
ChatMessage, ConversationDetail, ConversationSummary and their From<wire>
projections move from client-common/types.rs into api-model (new api::client
submodule — kept out of the crate root to avoid colliding with the wire-level
ConversationSummary). client-common::types re-exports them, so existing
client_common::{ChatMessage, ConversationDetail, ConversationSummary} paths are
unchanged. This lets the wasm client cores build/convert these views without
client-common's native tail.

api-model still compiles to wasm32; full workspace builds; fmt + clippy clean;
api-model/client-common tests pass.
