+++ crates/api-model/src/lib.rs
@@ -10,6 +10,7 @@ use std::collections::BTreeMap;
 
 use serde::{Deserialize, Serialize};
 
+pub mod client;
 pub mod signal;
 pub use signal::{SignalEvent, map_event_to_signal};
